// src/app/modules/cpassapi/model/subimpegno.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Ente } from './ente';
import { Fornitore } from './fornitore';
import { Impegno } from './impegno';


export interface Subimpegno { 
    id?: string;
    annoEsercizio?: number;
    anno?: number;
    numero?: number;
    descrizione?: string;
    stato?: string;
    importoAttuale?: number;
    importoIniziale?: number;
    liquidatoAnnoPrecedente?: number;
    numeroCapitolo?: number;
    numeroArticolo?: number;
    descrizioneCapitolo?: string;
    annoProvvedimento?: number;
    numeroProvvedimento?: number;
    settoreProvvedimento?: string;
    ente?: Ente;
    fornitore?: Fornitore;
    impegno?: Impegno;
    disponibile?: number;
    importo?: number;
}